Moved BboxPredictor to eval.py.
